refactor(backend): remove library import cycle
